REMARK FIXES AND DRAFT FIXES

+++ src/VectorFlow/Pages/MTA/MDM/ViewModify/useViewModify.ts
@@ -1592,7 +1592,11 @@ const useViewModify = (pageType:string) => {
 
       const onSaveToDraft = async () => {
         try{
-          dispatch(REMOVE_COLDEFS(['checkbox']))
+          const colDefs = ref.current?.api.getColumnDefs() || [];
+          const checkboxExists = colDefs.some((col: any) => col.field === "checkbox");
+          if (checkboxExists) {
+            dispatch(REMOVE_COLDEFS(["checkbox"]));
+          }
           dispatch(UPDATE_IS_SAVING_DRAFT(true))
           let newData:any = []
           const errorOrWarning = activeMaster.rowData.find((row:any)=>(Object.keys(row).includes('error'))||(Object.keys(row).includes('warning')));
@@ -1627,8 +1631,11 @@ const useViewModify = (pageType:string) => {
           }
           console.log(newData)
           const res = await postDraftChunks(newData)
+          if(checkboxExists){
+            addCheckBoxColDefs()
+          }
           if(res){
-            dispatch(UPDATE_PROGRESS_STATE("savedToDraft"))
+            //dispatch(UPDATE_PROGRESS_STATE("savedToDraft"))
             if(draftID.length > 0){
               return notifySuccess("Draft Updated Successfully")
             }
@@ -1637,15 +1644,15 @@ const useViewModify = (pageType:string) => {
             }
           }
           notifyError("Something Went Wrong")
-          if(activeMaster.progress==='uploaded'){
-            addCheckBoxColDefs()
-          }
+          // if(activeMaster.progress==='uploaded'){
+          //   addCheckBoxColDefs()
+          // }
           return false
         }catch(err){
           notifyError("Something Went Wrong")
-          if(activeMaster.progress==='uploaded'){
-            addCheckBoxColDefs()
-          }
+          // if(activeMaster.progress==='uploaded'){
+          //   addCheckBoxColDefs()
+          // }
           return false
         }finally{
           dispatch(UPDATE_IS_SAVING_DRAFT(false))
@@ -1657,7 +1664,16 @@ const useViewModify = (pageType:string) => {
         const currentMasterData = masters.find((master:MDMMasterState)=>master.id === activeMaster.id)
         console.log("CURRENTMATTER",currentMasterData?.rowData)
         if(currentMasterData) dispatch(UPDATE_ROW_DATA(currentMasterData.rowData))
-        dispatch(REMOVE_COLDEFS(['error','warning']));
+          const errorExist = currentMasterData?.rowData?.some((row: any) => row?.error !== undefined) || false;
+          const warningExist = currentMasterData?.rowData?.some((row: any) => row?.warning !== undefined) || false;
+        
+          if(!errorExist){
+            dispatch(REMOVE_COLDEFS(['error']));
+          }
+          if(!warningExist){
+            dispatch(REMOVE_COLDEFS(['warning']));
+          }
+        //dispatch(REMOVE_COLDEFS(['error','warning']));
         dispatch(UPDATE_PROGRESS_STATE('editOnline'));
         setEnableEditOnlineReset(false)
       }
